requirements.txt
ansi2html==1.6.0
asgiref==3.4.1
backcall==0.2.0
bidict==0.21.2
Brotli==1.0.9
certifi==2021.5.30
charset-normalizer==2.0.2
click==8.0.1
colorama==0.4.4
dash==1.21.0
dash-bootstrap-components==0.12.2
dash-core-components==1.17.1
dash-draggable==0.0.2
dash-extensions==0.0.58
dash-html-components==1.1.4
dash-oop-components==0.0.10
dash-renderer==1.9.1
dash-table==4.12.0
dash-treeview-antd==0.0.1
debugpy==1.3.0
decorator==5.0.9
dill==0.3.4
dpd-components==0.1.0
et-xmlfile==1.1.0
email_validator==1.1.2
Events==0.4
Flask==2.0.1
Flask-Caching==1.10.1
Flask-Compress==1.10.1
Flask-Login==0.5.0
Flask-SocketIO==5.1.0
flask_login==0.5.0
flask_migrate==2.7.0
flask_wtf==0.14.3
flask_sqlalchemy==2.4.4
future==0.18.2
idna==3.2
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyter-dash==0.4.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
more-itertools==8.8.0
numpy==1.21.0
openpyxl==3.0.7
oyaml==1.0
pandas==1.3.0
parso==0.8.2
pickleshare==0.7.5
plotly==5.1.0
prompt-toolkit==3.0.19
Pygments==2.9.0
python-dateutil==2.8.2
python-decouple==3.4
python-engineio==4.2.0
python-socketio==5.3.0
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
requests==2.26.0
retrying==1.3.3
scipy==1.7.0
shortuuid==1.0.1
six==1.16.0
sqlparse==0.4.1
tenacity==8.0.1
tornado==6.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
Werkzeug==2.0.1
SQLAlchemy==1.3.23
pysqlite3
rq==1.9.0
gunicorn

WTForms~=2.3.3
redis~=3.5.3
setuptools~=57.4.0